fix: pin ruff so CI, pre-commit, and local agree
CI installs an unpinned ruff, so it resolved to 0.16.2 while
.pre-commit-config.yaml pins v0.15.20 and lint passed locally. The two versions
disagree irreconcilably here: 0.16 stopped enabling E402 by default, which turns
the noqa waivers 0.15 requires in tests/test_outer_loop.py into RUF100
"unused noqa" errors. No noqa state satisfies both, so the float itself is the
bug.

Pin ruff==0.15.20 in the dev extra to match the pre-commit rev, and fix the real
findings 0.16.2 surfaced (all of which 0.15.20 also accepts):

- I001: import blocks left un-sorted by the earlier mechanical sampling ->
  circuit_run rewrites, in 4 files.
- C408: dict() calls rewritten as literals in test_circuit_generator.py.
- RUF046: drop a redundant int() around round(), which already returns int.

Bump the pin and the pre-commit rev together in future.

+++ src/embasi_qiskit_integration/circuit_run/base.py
@@ -86,7 +86,8 @@ class MockSampler(SamplerMixin):
         # total exactly by handing any rounding remainder to the largest bin.
         scaled: dict[str, int] = {}
         for bitstring, count in self._counts.items():
-            scaled[bitstring] = int(round(count * shots / total))
+            # round() on a float already returns an int; no cast needed.
+            scaled[bitstring] = round(count * shots / total)
         drift = shots - sum(scaled.values())
         if drift != 0 and scaled:
             top = max(scaled, key=scaled.__getitem__)
